Fixed minor annoyances
Added the ability to specify which port to listen on, as well as which
port to connect to.  Also made it optional to start listening, rather
than the server end trying to listen on a hard coded port from the
start.
I plan to make it prettier in the next commit, but hope this helps.